Standardize code formatting
Code formatting was not good. Indentation was wrong, as spaces got mixed
with tabs, and it was also not following any standard.

This commit will standardize code formatting, following VSCode default
formatter. There were a couple settings changed as described below.

Notable standards that have been set:
- <head> and <body> are not indented (default VSCode setting)
- CSS Rules opens braces in new line
- JavaScript opens block and function braces in new line
- All indentations are tabs

I am aware of JavaScript automatic semicolon placement, and how that
interacts with 'return' when braces are put in a new line. However, at
this point it feels easier to write cleaner code using braces in a new
line. Either way, this rule may change in the future, if needed.