Make pokemon from list be selectable and saved in the route box
The pokemon list was only listing the existing pokemon.

Now, when a pokemon in the list is clicked, it is selected and its name
shows in the text input field.

A placeholder was also added to indicate the input should be used as a
search mechanism for pokemon and to make switching between pokemon in a
route box smoother.

The implementation was using the focus and blur events in the input
field and click event in the list item.

Notice that, while the functionality is working very well, the
implementation logic is a bit hard to grasp and may be improved in the
future.