Make selected pokemon image show in left circle in route box
Whenever a pokemon was selected from the list, it only appeared in the
text input field.

Now its image will also show inside the left circle.

The implementation was done by copying the image source of the image of
the list item that was clicked to a new image element appended to the
left circle.

Considerations:
- So far, only small pokemon tend to be beautifully presented inside the
  circle
- Big pokemon will extrapole the circle's border (it kind of creates a
  nice effect, but I'm not sure it works well for all pokemon)
- The image rendering was set to pixelated, because when the image size
  was increased, it didn't look good

In the future, a better implementation considering pokemon sizes may be
put in place.