Fix pokemon image in pokemon list being valid for pokemon selection
The pokemon images in the pokemon list are implemented to overflow their
list item. That way big pokemon are not bound by its list item.

However, since the image can overflow the list item, it should not be
valid for pokemon selection, as you could click in a pokemon and select
the one below it in the list.

The fix was implemented using the CSS declaration `pointer-events:
none`, applied to all children of the part of list item that is meant to
select pokemon on click. This way, list items can include many elements
and none of them will trigger the click event.

A note about the implementation: all browsers from 2013 forwards can use
this CSS declaration, except Opera Mini. It's still unclear the level of
browser support the site shall provide, but it feels ok to use something
that is available in all browsers of the past decade.