Refactor script.js to turn route box into a class
The route box, previously not defined formally in the code, was just a
function that essentially put together all elements it needed.

The function was starting to become big and it did a lot of different
things. Also, referencing certain elements was starting to become
questionable in the implementation.

Now route box has been turned to its own class. The way it was done was
by using the Factory Design Pattern, in which you have the
`route_box_factory` constant, that can be used like a constructor for
new route box objects.

The implementation of the factory may not be perfect, as it's the first
time this design pattern has been implemented in the project.