Add scrollbar to pokemon list
The pokemon list did not have limited size and scrollbar. It was limited
by setting the number of list items displayed to a small number.

Now the pokemon list has a scrollbar. The implementation was only of a
vertical scrollbar, limiting the size of the list to 8 list items and
using the `overflow-y` set to `auto` as a CSS rule for `.pokemon-list`.

It remains necessary to style the newly added scrollbar and to optimize
the creation of the pokemon list.