Add search functionality to input field
The input field had no functionality yet.

Now, when typing something in the input field, if the input is a
substring of a pokemon's name, that pokemon will be included in the
pokemon list, otherwise it won't.

Big warning: this is a very primitive implementation, and it's really
slow.

The complexity of the search algorithm is O(number of pokemon * length
of pokemon name). Since currently there are 898 pokemon, and the longest
pokemon name has 12 characters, the amount of operations is about 11000.
This amount of operations feels ok, but the complexity can probably be
improved.

The current problems with the implementation are the following (note
that the times were measured locally):
- The generation of the list is taking about 3 seconds
- The empty string search (which basically includes all pokemon in the
  list) takes about 100 milliseconds