Change pokemon selection wrapper to absolute position
The pokemon selection wrapper, which wraps both the pokemon input field
and the pokemon list previously did not have any CSS rules that altered
its position.

Now it's set to absolute position. The design of the route box was kept
the same, by changing the rules to maintain the same structure.

This fixes the bug of clicking above the pokemon input field and turning
its text color to placeholder. It was a side effect of the pokemon
selection wrapper position in the route box being actually above the
pokemon input field.

Notes:
- Before this commit, it was presented in the code as
  `pokemon_select_wrapper` and its CSS class as `pokemon-select-wrapper`.
  "select" was transformed to "selection", to better describe it.
- A possible better solution would be to remove this wrapper entirely
  and pass its CSS rule as a new class to pokemon input field and
  pokemon list. The problem is that a funcionality requires a CSS rule
  in which the pokemon list is active and it references the pokemon
  input field, which is a sibling that comes BEFORE it on the HTML
  structure. Unfortuanally, there's no straightforward way to do this in
  CSS.