Enhance pokemon search by sorting the list based on match quality
When the user searched for a pokemon, all pokemon with the user input as
a substring were shown in the list, ordered by their pokemon id.

Now the resulting list will be ordered by the quality of the match of
the input versus the pokemon name. In other words, it tries to consider
the intent of the user during the search.

The implementation of the sorting considers the following:
  - It's probable that the user will begin its search by the first
    letter of the intended pokemon name
  - It's probable that the input will contain as long of a contiguous
    substring of the intended pokemon name as possible
  - It's likely that the user will search for earlier characters of the
    intended pokemon name

For optimization:
  - A flag was set so the list will only sort and make all items visible
    after a non empty search
  - A function that detaches an element in the DOM for later usage was
    built