Add lazy loading to pokemon list images
When the list was created, the assignment of all image sources at once
caused the page to be stuck for a while.

Now images are only loaded when either the user scrolls over them or
nothing else is happening.

The implementation was done using the `loading: "lazy"` HTML attribute.
This makes the images only render when they show up for the first time
in the page.

In addition to the above, `window.requestIdleCallback` was also used to
do the opposite: removing the `loading` attribute from the images, only
during idle moments. This will effectively render the images even before
the user scrolls over them, so that when they actually show up, they are
already rendered.

Only the `window.requestIdleCallback` was not sufficient, because when
the user is actually scrolling through the list, the site is not idle
and therefore unloaded images will not load. Only the `loading` would
work; the second part is only an enhancement.

Another implementation using the Intersection Observer API was tried. It
also worked, but it seems that `loading` renders images quicker upon
scrolling. This implementation was left commented out in this commit.