Remove spellchecking from pokemon input field
When the user inserted a wrongly spelled word inside the input field,
after a bit it showed the red wavy line below it.

Now it does not. It's undesired because when searching for pokemon, the
input search is most likely not a correctly spelled word.

It was implemented setting the input field HTML attribute `spellcheck`
to false.