Add feedback to mouse hover and mouse click to pokemon list items
When the user hovered over the list, no feedback was given on the
current selectable pokemon.

Now, when the user hovers over a list item, it will be darker. While the
user clicks and holds a list item, it will remain darker and no other
list items will be darker.

The implementation was done with simple CSS rules, using specially the
`hover` and `active` pseudoclasses.

For optimization, the `click_area` element, child of pokemon list item,
was removed, since it was not necessary. The list item itself can
already serve click area's purpose: defining the click area to select a
pokemon from the list.