Make route title renamable on click
The route title simply showed the name of the route.

Now when the user clicks in the route title, they are able to edit it.

The implementation was done using the `contentEditable` HTML attribute.

For consistency, in this commit all `setAttribute` were changed to
`.attribute = `. They are quite similar, but it seems it's better to
work with object properties directly.