The "placeholder" and "value" colors are now set to CSS variables
For context, there are two colors used a lot throughout the site: one is
called the "placeholder" color and the other the "value" color.

Their names come from the color present in the pokemon input field when
it shows a placeholder and a value, respectively. Since they are used in
many places, it feels ok to set them to a CSS variable.

The implementation was done using the `:root` selector and including
inside it `--placeholder-color` and `--value-color` declarations, which
are used in other rules by the `var()` function.

Also, a small typo was corrected in the main page about the name of the
website.