Change state dropdown to same style as pokemon list
State dropdown was a `select` HTML attribute. It worked well, but it's
harder to insert images into it and also, in some browsers, it looked
very different from the rest of the site.

Now it works similarly as the pokemon list, using basically the same
HTML elements and CSS rules. It's easier for customization and keeps the
default look of the website.